feat: extract exercise sorting functionality and integrate usage data across multiple screens
